+++ src/pages/ManagePosts/ManagePosts.jsx
@@ -39,18 +39,38 @@ function ManagePosts() {
     navigate(`/forms/add-post/${propertyId}`);
   };
 
+  // const handleDeleteClick = async (id) => {
+  //   try {
+  //     const response = await axios.delete(FETCH_ALL_POSTS + `/${id}`, {
+  //       withCredentials: true,
+  //     });
+  //     if (response.data.success) {
+  //       setProperties(properties.filter((property) => property.id !== id));
+  //     } else {
+  //       console.error("Error deleting property:", response.data.message);
+  //     }
+  //   } catch (error) {
+  //     console.error("Error deleting property:", error);
+  //   }
+  // };
+
   const handleDeleteClick = async (id) => {
-    try {
-      const response = await axios.delete(FETCH_ALL_POSTS + `/${id}`, {
-        withCredentials: true,
-      });
-      if (response.data.success) {
-        setProperties(properties.filter((property) => property.id !== id));
-      } else {
-        console.error("Error deleting property:", response.data.message);
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this post?"
+    );
+    if (confirmDelete) {
+      try {
+        const response = await axios.delete(FETCH_ALL_POSTS + `/${id}`, {
+          withCredentials: true,
+        });
+        if (response.data.success) {
+          setProperties(properties.filter((property) => property.id !== id));
+        } else {
+          console.error("Error deleting post:", response.data.message);
+        }
+      } catch (error) {
+        console.error("Error deleting post:", error);
       }
-    } catch (error) {
-      console.error("Error deleting property:", error);
     }
   };
